+++ cypress/cypress/e2e/Q-practice-quiz-workflow.cy.ts
@@ -1036,6 +1036,8 @@ describe('Different practice quiz workflows', function () {
       `[data-cy="publish-practice-quiz-${this.data.scheduled.name}"]`
     ).click()
 
+    const courseStartMonthDelta = -12 - new Date().getMonth()
+
     // check that if publication date is before course start date, submission is disabled
     cy.get('[data-cy="schedule-practice-quiz-publication"]').should(
       'be.disabled'
@@ -1044,13 +1046,14 @@ describe('Different practice quiz workflows', function () {
       cyString: 'practice-quiz-available-from',
       deselectorString: 'publish-immediately-header',
       datetime: {
-        monthDelta: -36,
-        day: 15,
-        hour: 12,
+        monthDelta: courseStartMonthDelta,
+        day: 1,
+        hour: 0,
         minute: 0,
-        validation: getDatetimeValidationString(-36, '15') + ', 12:00',
+        validation:
+          getDatetimeValidationString(courseStartMonthDelta, '01') + ', 00:00',
       },
-    }) // select publication date 3 years in the past -> course start date is beginning of the previous year
+    }) // select publication date equal to course start date (must be strictly after)
     cy.get('[data-cy="schedule-practice-quiz-publication"]').should(
       'be.disabled'
     )
@@ -1060,7 +1063,7 @@ describe('Different practice quiz workflows', function () {
       cyString: 'practice-quiz-available-from',
       deselectorString: 'publish-immediately-header',
       datetime: {
-        monthDelta: 40,
+        monthDelta: 4 - courseStartMonthDelta,
         day: 15,
         hour: 12,
         minute: 0,